Refactor matcher
- Use #[default]
- Remove useless code
