+++ crates/matcher/src/lib.rs
@@ -38,305 +38,21 @@
 
 mod algo;
 mod bonus;
+mod matchers;
 #[cfg(test)]
 mod tests;
 
-pub use self::algo::{fzy, skim, substring, FuzzyAlgorithm};
+pub use self::algo::FuzzyAlgorithm;
 pub use self::bonus::cwd::Cwd;
 pub use self::bonus::language::Language;
 pub use self::bonus::Bonus;
-use crate::substring::substr_indices;
-use grep_regex::{RegexMatcher, RegexMatcherBuilder};
+pub use self::matchers::{BonusMatcher, ExactMatcher, FuzzyMatcher, InverseMatcher, WordMatcher};
 use std::path::Path;
 use std::sync::Arc;
-use types::{
-    CaseMatching, ClapItem, ExactTerm, ExactTermType, FuzzyTerm, FuzzyText, InverseTerm,
-    MatchedItem, WordTerm,
-};
+use types::{CaseMatching, ClapItem, FuzzyText, MatchedItem};
 
 // Re-export types
 pub use types::{MatchResult, MatchScope, Query, Score};
-
-#[derive(Debug, Clone, Default)]
-pub struct InverseMatcher {
-    inverse_terms: Vec<InverseTerm>,
-}
-
-impl InverseMatcher {
-    pub fn new(inverse_terms: Vec<InverseTerm>) -> Self {
-        Self { inverse_terms }
-    }
-
-    pub fn inverse_terms(&self) -> &[InverseTerm] {
-        &self.inverse_terms
-    }
-
-    /// Returns `true` if any inverse matching is satisfied, which means the item should be
-    /// ignored.
-    pub fn match_any(&self, match_text: &str) -> bool {
-        self.inverse_terms
-            .iter()
-            .any(|inverse_term| inverse_term.exact_matched(match_text))
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct ExactMatcher {
-    exact_terms: Vec<ExactTerm>,
-    case_matching: CaseMatching,
-}
-
-impl ExactMatcher {
-    pub fn new(exact_terms: Vec<ExactTerm>, case_matching: CaseMatching) -> Self {
-        Self {
-            exact_terms,
-            case_matching,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.exact_terms.is_empty()
-    }
-
-    pub fn exact_terms(&self) -> &[ExactTerm] {
-        &self.exact_terms
-    }
-
-    /// Returns an optional tuple of (score, indices) if all the exact searching terms are satisfied.
-    pub fn find_matches(&self, full_search_line: &str) -> Option<(Score, Vec<usize>)> {
-        let mut indices = Vec::<usize>::new();
-        let mut exact_score = Score::default();
-
-        if full_search_line.is_empty() {
-            return None;
-        }
-
-        for term in &self.exact_terms {
-            let sub_query = &term.text;
-
-            match term.ty {
-                ExactTermType::Exact => {
-                    if let Some((score, sub_indices)) =
-                        substr_indices(full_search_line, sub_query, self.case_matching)
-                    {
-                        indices.extend_from_slice(&sub_indices);
-                        exact_score += score.max(sub_query.len() as Score);
-                    } else {
-                        return None;
-                    }
-                }
-                ExactTermType::PrefixExact => {
-                    let trimmed = full_search_line.trim_start();
-                    let white_space_len = full_search_line.len().saturating_sub(trimmed.len());
-                    if trimmed.starts_with(sub_query) {
-                        let mut match_start = -1i32 + white_space_len as i32;
-                        let new_len = indices.len() + sub_query.len();
-                        indices.resize_with(new_len, || {
-                            match_start += 1;
-                            match_start as usize
-                        });
-                        exact_score += sub_query.len() as Score;
-                    } else {
-                        return None;
-                    }
-                }
-                ExactTermType::SuffixExact => {
-                    let total_len = full_search_line.len();
-                    let trimmed = full_search_line.trim_end();
-                    let white_space_len = total_len.saturating_sub(trimmed.len());
-                    if trimmed.ends_with(sub_query) {
-                        // In case of underflow, we use i32 here.
-                        let mut match_start = total_len as i32
-                            - sub_query.len() as i32
-                            - 1i32
-                            - white_space_len as i32;
-                        let new_len = indices.len() + sub_query.len();
-                        indices.resize_with(new_len, || {
-                            match_start += 1;
-                            match_start as usize
-                        });
-                        exact_score += sub_query.len() as Score;
-                    } else {
-                        return None;
-                    }
-                }
-            }
-        }
-
-        // Add an exact search term bonus whether the exact matches exist or not.
-        //
-        // The shorter search line has a higher score.
-        exact_score += (512 / full_search_line.len()) as Score;
-
-        Some((exact_score, indices))
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct FuzzyMatcher {
-    match_scope: MatchScope,
-    fuzzy_algo: FuzzyAlgorithm,
-    fuzzy_terms: Vec<FuzzyTerm>,
-    case_matching: CaseMatching,
-}
-
-impl FuzzyMatcher {
-    pub fn new(
-        match_scope: MatchScope,
-        fuzzy_algo: FuzzyAlgorithm,
-        fuzzy_terms: Vec<FuzzyTerm>,
-        case_matching: CaseMatching,
-    ) -> Self {
-        Self {
-            match_scope,
-            fuzzy_algo,
-            fuzzy_terms,
-            case_matching,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.fuzzy_terms.is_empty()
-    }
-
-    pub fn find_matches(&self, item: &Arc<dyn ClapItem>) -> Option<(Score, Vec<usize>)> {
-        item.fuzzy_text(self.match_scope)
-            .as_ref()
-            .and_then(|fuzzy_text| self.match_fuzzy_text(fuzzy_text))
-    }
-
-    pub fn match_fuzzy_text(&self, fuzzy_text: &FuzzyText) -> Option<(Score, Vec<usize>)> {
-        let fuzzy_len = self.fuzzy_terms.iter().map(|f| f.len()).sum();
-
-        // Try the fuzzy terms against the matched text.
-        let mut fuzzy_indices = Vec::with_capacity(fuzzy_len);
-        let mut fuzzy_score = Score::default();
-
-        for term in self.fuzzy_terms.iter() {
-            let query = &term.text;
-            if let Some(MatchResult { score, indices }) =
-                self.fuzzy_algo
-                    .fuzzy_match(query, fuzzy_text, self.case_matching)
-            {
-                fuzzy_score += score;
-                fuzzy_indices.extend(indices);
-            } else {
-                return None;
-            }
-        }
-
-        Some((fuzzy_score, fuzzy_indices))
-    }
-}
-
-/// Word matching using the `RegexMatcher`.
-#[derive(Debug, Clone, Default)]
-pub struct WordMatcher {
-    matchers: Vec<(WordTerm, RegexMatcher)>,
-}
-
-impl WordMatcher {
-    pub fn new(word_terms: Vec<WordTerm>) -> Self {
-        let matchers = word_terms
-            .into_iter()
-            .filter_map(|word_term| {
-                RegexMatcherBuilder::default()
-                    .word(true)
-                    .build(&word_term.text)
-                    .ok()
-                    .map(|word_matcher| (word_term, word_matcher))
-            })
-            .collect();
-
-        Self { matchers }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.matchers.is_empty()
-    }
-
-    pub fn find_matches(&self, line: &str) -> Option<(Score, Vec<usize>)> {
-        use grep_matcher::Matcher;
-
-        let mut score = Score::default();
-
-        let byte_indices: Vec<_> = self
-            .matchers
-            .iter()
-            .filter_map(|(word_term, word_matcher)| {
-                word_matcher
-                    .find_at(line.as_bytes(), 0)
-                    .ok()
-                    .flatten()
-                    .map(|mat| {
-                        let start = mat.start();
-                        let end = mat.end();
-                        score += word_term.score(start);
-                        start..end
-                    })
-            })
-            .flatten()
-            .collect();
-
-        // In order to be consistent with the other matchers which use char-positions, even all
-        // char-positions will be converted to byte-positions before sending to Vim/Neovim in the end.
-        let indices = line
-            .char_indices()
-            .enumerate()
-            .filter_map(|(char_idx, (byte_idx, _char))| {
-                if byte_indices.contains(&byte_idx) {
-                    Some(char_idx)
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>();
-
-        if !indices.is_empty() {
-            Some((score, indices))
-        } else {
-            None
-        }
-    }
-}
-
-/// [`BonusMatcher`] only tweaks the match score.
-#[derive(Debug, Clone, Default)]
-pub struct BonusMatcher {
-    bonuses: Vec<Bonus>,
-}
-
-impl BonusMatcher {
-    pub fn new(bonuses: Vec<Bonus>) -> Self {
-        Self { bonuses }
-    }
-
-    /// Returns the sum of bonus score.
-    pub fn calc_item_bonus(
-        &self,
-        item: &Arc<dyn ClapItem>,
-        base_score: Score,
-        base_indices: &[usize],
-    ) -> Score {
-        self.bonuses
-            .iter()
-            .map(|b| b.item_bonus_score(item, base_score, base_indices))
-            .sum()
-    }
-
-    /// Returns the sum of bonus score.
-    pub fn calc_text_bonus(
-        &self,
-        bonus_text: &str,
-        base_score: Score,
-        base_indices: &[usize],
-    ) -> Score {
-        self.bonuses
-            .iter()
-            .map(|b| b.text_bonus_score(bonus_text, base_score, base_indices))
-            .sum()
-    }
-}
 
 #[derive(Debug, Clone, Default)]
 pub struct MatcherBuilder {
