+++ crates/matcher/src/matchers/mod.rs
@@ -1,0 +1,11 @@
+mod bonus_matcher;
+mod exact_matcher;
+mod fuzzy_matcher;
+mod inverse_matcher;
+mod word_matcher;
+
+pub use self::bonus_matcher::BonusMatcher;
+pub use self::exact_matcher::ExactMatcher;
+pub use self::fuzzy_matcher::FuzzyMatcher;
+pub use self::inverse_matcher::InverseMatcher;
+pub use self::word_matcher::WordMatcher;
