+++ crates/matcher/src/tests.rs
@@ -226,3 +226,22 @@ fn test_word_matcher() {
             .collect::<String>()
     );
 }
+
+#[test]
+fn test_rank() {
+    let items = vec![
+        Arc::new("pythonx/clap/fuzzymatch-rs/.cargo/config"),
+        Arc::new("crates/maple_core/src/config.rs"),
+        Arc::new("config.toml"),
+        Arc::new(".editorconfig"),
+    ];
+
+    let query: Query = "config".into();
+    let matcher = MatcherBuilder::new().build(query);
+
+    for item in items {
+        let matched_item = matcher.match_item(item).unwrap();
+
+        println!("{matched_item:?}");
+    }
+}
